requirements.txt
langchain==0.1.13
python-dotenv==1.0.0
streamlit==1.32.2
google-generativeai==0.4.1
unstructured==0.12.6
tiktoken==0.4.0
faiss-cpu==1.8.0
python-dotenv==1.0.1
libmagic==1.0
python-magic==0.4.27
python-magic-bin==0.4.14